requirements.txt
websocket-client>=1.6.4
numpy>=1.26.0,<2.0.0
tensorflow==2.16.1
scikit-learn>=1.4.0
joblib>=1.3.2
openai>=1.61.0
pandas>=2.2.3
plotly>=6.0.0
streamlit>=1.41.1
yfinance>=0.2.52
beautifulsoup4>=4.12.3
html5lib>=1.1
lxml>=5.1.0
requests>=2.31.0 